Elasticsearch always returns a GUID, while Solr does not for now.  Needed to fix-up the integration tests due to this.
